[revm] Move Gas, and machine outside
