+++ crates/revm/src/interpreter/mod.rs
@@ -4,8 +4,10 @@ mod machine;
 mod contract;
 pub(crate) mod memory;
 mod stack;
+mod gas;
 
 pub use contract::Contract;
 pub use machine::*;
 pub use memory::Memory;
 pub use stack::Stack;
+pub use gas::Gas;
